requirements.txt
fastapi -
fastapi "python-jose[cryptography]" 
uvicorn[standard] -
gunicorn
bcrypt -
python-dotenv -
mysql-connector-python -
python-jose[cryptography]
aiofiles              # ← נדרש להעלאת קבצים
pillow -                # ← אם יש טיפול בתמונות 
httpx                 # ← אם אתה שולח בקשות לשרתים חיצוניים
torchvision -
diffusers -
transformers -
torch - 
accelerate


